backend/requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
flask-cors>=4.0.0
gunicorn>=20.1.0

alembic>=1.11.0

python-dotenv>=1.0.0

pyjwt>=2.7.0
cryptography>=41.0.0

python-dateutil>=2.8.0
psycopg2-binary>=2.9.0